% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amanida.R
\docType{package}
\name{amanida}
\alias{amanida}
\title{amanida}
\description{
amanida: A package for Adapted Meta-Analysis with non-integral data
}
\author{
Maria Llambrich and Eudald Correig

Adaptation of results combination for meta-analysis using only significance and effect size.
\itemize{
\item P-values and fold-change are combined to obtain a global significance on each metabolite.
\item Produces a volcano plot summarizing the relevant results from meta-analysis.
\item Vote-counting reports for metabolites
}
}
