% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amanida_read.R
\name{amanida_read}
\alias{amanida_read}
\title{Import data}
\usage{
amanida_read(file, coln, separator = NULL)
}
\arguments{
\item{file}{path to file}

\item{coln}{columns names to use}

\item{separator}{the separator used on file}
}
\value{
tibble table with data imported
}
\description{
\code{amanida_read} imports the data and formats for metamet function
}
\details{
Note that \code{amanida_read} skips rows with missing values or NA.

Formats compatible are csv, xlsx, xls or txt.
}
\examples{
coln = c("Compound Name", "P-value", "Fold-change", "N total", "References")
input_file <- system.file("extdata", "dataset2.csv", package = "amanida")
datafile <- amanida_read(input_file, coln, separator=";")

}
