% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amanida_vote.R
\name{amanida_vote}
\alias{amanida_vote}
\title{Import data}
\usage{
amanida_vote(datafile)
}
\arguments{
\item{file}{datafile data imported using amanida_read function}
}
\value{
METAtable S4 object with vote-counting for each compound on @slot vote
}
\description{
\code{amanida_vote} performs vote-counting on data
}
\details{
Note that \code{amanida_vote} skips rows with missing values or NA.

Formats compatible are csv, xlsx, xls or txt.
}
\examples{
coln = c("Compound Name", "Behaviour", "References")
input_file <- system.file("extdata", "dataset2.csv", package = "amanida")

data_votes <- amanida_read(input_file, coln, separator = ";")

vote_result <- amanida_vote(data_votes)

}
