% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croms.R
\docType{package}
\name{croms}
\alias{croms}
\title{croms}
\description{
croms: A package for Combining results of metabolomics studies
}
\author{
Maria Llambrich and Eudald Correig

Statistical results combination of metabolomics studies using a meta-analysis approach.
\itemize{
\item P-values and fold-change are combined to obtain a global significance on each metabolite.
\item Produces a volcano plot summarizing the relevant results from meta-analysis.
\item Vote-counting reports for metabolites
}
}
