% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{explore_plot}
\alias{explore_plot}
\title{Plot for compounds divergence in reports}
\usage{
explore_plot(data)
}
\arguments{
\item{data}{an tibble obtained by \code{amanida_read}}
}
\value{
a ggplot bar-plot showing the sum of votes for each compound divided by the trend
}
\description{
\code{explore_plot} creates a bar-plot showing the votes divided in up-regulated and down-regulated and the global result for each compound.
}
\details{
Sum of votes divided by trend are plotted, then is obtained the total result by compound summing both trends.
}
\examples{
data("sample_data")

explore_plot(sample_data)

}
