% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaplot}
\alias{metaplot}
\title{Volcano plot of combined results}
\usage{
metaplot(mets, cutoff = NULL)
}
\description{
\code{metaplot} returns a volcano plot of the combined results on each metabolite obtained by metmet function
}
\details{
Results are presented as -log10 for p-value and log2 for fold-change.
Values over the cut off are labeled. If not cutoff is provided will be used alpha 0.05 for p-value and 1.5 for logarithmic fold-change.

@param mets an S4 METAtable object
@param cutoff values for p-value and fold-change significance

@return plot of results

@example
coln = c("Compound Name", "P-value", "Fold-change", "N total", "References")
datafile <- data.read("~/OneDrive - URV/metaanalysis_pack/dataset2.xlsx", coln)
res.met <- metmet(datafile)
metaplot(res.met, c(0.01, 2))

@import dplyr, ggplot2
}
