% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{METAtables-class}
\alias{METAtables-class}
\alias{.METAtables}
\title{An S4 class to return results from metamet function}
\description{
An S4 class to return results from metamet function
}
\section{Slots}{

\describe{
\item{\code{stat}}{results for statistics combining p-values and fold-changes}

\item{\code{vote}}{vote-counting for metabolites}
}}

