% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example input data for the amanida function}
\format{
A data frame with 93 rows and 6 variables:
\describe{
\item{Compound name}{Name of the compound under study}
\item{Behaviour}{UP, DOWN}
\item{N total}{Number of samples of the compound}
\item{P-value}{P-value}
\item{Fold-change}{Fold-change}
\item{References}{References}
}
}
\usage{
sample_data
}
\description{
A dataset containing results from meta-analysis of metabolomic studies
}
\keyword{datasets}
