% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vote_plot}
\alias{vote_plot}
\title{Bar-plot for compounds number of reports}
\usage{
vote_plot(mets)
}
\arguments{
\item{mets}{an S4 METAmet object obtained by \code{metmet}}
}
\value{
a ggplot bar-plot showing the number of articles per compound
}
\description{
\code{vote_plot} creates a bar-plot showing the number of entries for each compound.
}
\details{
Independently of the compound trend, the total number of entries on each compound are plotted.
}
\examples{
data("sample_data")

result <- compute_amanida(sample_data)
vote_plot(result)

}
