% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{voteplot}
\alias{voteplot}
\title{Bar-plot for compouends number of reports}
\usage{
voteplot(mets)
}
\description{
\code{voteplot} creates a bar-plot showing the number of entries for each compound.
}
\details{
Independently of the compound trend, the total number of entries on each compound are plotted.

@params mets an S4 METAmet object obtained by \code{metmet}

@return bar-plot of results
@example
coln = c("Compound Name", "P-value", "Fold-change", "N total", "References")
datafile <- data.read("~/OneDrive - URV/metaanalysis_pack/dataset2.xlsx", coln)
res.met <- metmet(datafile)
voteplot(res.met)
}
