% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amanida.R
\docType{package}
\name{amanida}
\alias{amanida}
\alias{amanida-package}
\title{amanida}
\description{
amanida: A package for Meta-Analysis with non-integral data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mariallr/amanida}
}

}
\author{
Maria Llambrich, Eudald Correig and Raquel Cumeras

Results combination for meta-analysis using only significance and effect size.
\itemize{
\item P-values and fold-change are combined to obtain a global significance on each metabolite.
\item Produces a volcano plot summarizing the relevant results from meta-analysis.
\item Qualitative meta-analysis for metabolites
\item Graphical representation of qualitative analysis by bar plot
\item Trend explore plot to detect discrepancies between studies at a first glance
}
}
