% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data.read}
\alias{data.read}
\title{Import data}
\usage{
data.read(file, coln, separator = NULL)
}
\arguments{
\item{file}{path to file}

\item{coln}{columns names to use}

\item{separator}{the separator used on file}
}
\value{
tibble table with data imported
}
\description{
\code{data.read} imports the data and formats for metamet function
}
\details{
Note that \code{data.read} skips rows with missing values or NA.

Formats compatible are csv, xlsx, xls or txt.
}
\examples{

coln = c("Compound Name", "P-value", "Fold-change", "N total", "References")
data.read("~/OneDrive - URV/metaanalysis_pack/dataset2.xlsx", coln)

}
