% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R, R/plots.R
\name{metaplot}
\alias{metaplot}
\title{Volcano plot of combined results}
\usage{
metaplot(mets, cutoff = NULL)

metaplot(mets, cutoff = NULL)
}
\arguments{
\item{mets}{an S4 METAtable object}

\item{cutoff}{values for p-value and fold-change significance}
}
\value{
plot of results
}
\description{
\code{metaplot} returns a volcano plot of the combined results on each metabolite obtained by metmet function

\code{metaplot} returns a volcano plot of the combined results on each metabolite obtained by metmet function
}
\details{
Results are presented as -log10 for p-value and log2 for fold-change.
Values over the cut off are labeled. If not cutoff is provided will be used alpha 0.05 for p-value and 1.5 for logarithmic fold-change.

Results are presented as -log10 for p-value and log2 for fold-change.
Values over the cut off are labeled. If not cutoff is provided will be used alpha 0.05 for p-value and 1.5 for logarithmic fold-change.

@param mets an S4 METAtable object
@param cutoff values for p-value and fold-change significance

@return plot of results
}
