% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example input data for the amanida function}
\format{
A data frame with 143 rows and 6 variables:
\describe{
\item{id}{Name of the compound under study}
\item{pvalue}{P-value}
\item{foldchange}{Fold-change}
\item{N}{Number of samples of the compound}
\item{ref}{References}
\item{trend}{Trend: 1 (up), -1 (down) or 0 (none)}
}
}
\usage{
sample_data
}
\description{
A dataset containing results from meta-analysis of metabolomic studies
}
\keyword{datasets}
