% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vote_plot}
\alias{vote_plot}
\title{Bar-plot for compounds vote-counting}
\usage{
vote_plot(mets)
}
\arguments{
\item{mets}{an S4 METAmet object obtained by \code{compute_amanida}}
}
\value{
a ggplot bar-plot showing the vote-count per compound
}
\description{
\code{vote_plot} creates a bar-plot showing the vote-count for each compound.
}
\details{
Vote-couting is the sum of number of reports up-regulated and the substraction of reports down-regulated.
}
\examples{
data("sample_data")

result <- compute_amanida(sample_data)
vote_plot(result)

}
