% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{voteplot}
\alias{voteplot}
\title{Bar-plot for compounds number of reports}
\usage{
voteplot(mets)
}
\arguments{
\item{mets}{an S4 METAmet object obtained by \code{metmet}}
}
\value{
bar-plot of results
}
\description{
\code{voteplot} creates a bar-plot showing the number of entries for each compound.
}
\details{
Independently of the compound trend, the total number of entries on each compound are plotted.
}
\examples{
data("sample_data")

met <- metmet(sample_data)
voteplot(met)

}
